Movement 2.0
This is a full rehash of the game. There is movement (somewhat jerky at higher speeds), Jump with accellerated falling and ground detection, shooting with placeholder bullets and sounds, and aiming. The game has a working main menu and the core work for a pause menu in-game, but not implimented.